refactor: cost and quantity data types to double for increased precision
Updated various classes and services to change cost and quantity fields from long/int to double. This change enhances precision, especially for financial calculations and inventory management. Modified database schema to reflect these changes by altering column data types to double where necessary. Introduced a new unit field for parts to support more detailed inventory tracking. Updated frontend components and utility functions to handle and format these changes appropriately.

+++ api/src/main/java/com/grash/controller/PurchaseOrderController.java
@@ -105,8 +105,8 @@ public class PurchaseOrderController {
                             user.getCompany().getId());
             workflows.forEach(workflow -> workflowService.runPurchaseOrder(workflow, savedPurchaseOrder));
             PurchaseOrderShowDTO result = setPartQuantities(purchaseOrderMapper.toShowDto(savedPurchaseOrder));
-            long cost =
-                    result.getPartQuantities().stream().mapToLong(partQuantityShowDTO -> partQuantityShowDTO.getQuantity() * partQuantityShowDTO.getPart().getCost()).sum();
+            double cost =
+                    result.getPartQuantities().stream().mapToDouble(partQuantityShowDTO -> partQuantityShowDTO.getQuantity() * partQuantityShowDTO.getPart().getCost()).sum();
             String title = messageSource.getMessage("new_po", null, Helper.getLocale(user));
             String message = messageSource.getMessage("notification_new_po_request", new Object[]{result.getName(),
                             cost,
