+++ api/src/main/java/com/grash/dto/analytics/workOrders/WOCostsByDate.java
@@ -12,8 +12,8 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 public class WOCostsByDate {
-    private long partCost;
-    private long additionalCost;
-    private long laborCost;
+    private double partCost;
+    private double additionalCost;
+    private double laborCost;
     private Date date;
 }
